all tertiary group related files have been consolidated functionally into 'groupManagament'
